+++ application/controllers/PartController.php
@@ -109,33 +109,18 @@ class PartController extends Application
 	 */
 	public function build()
 	{
-		$tokens = array();
-		$properties = $this->properties->tail();
-		if (sizeof($properties) == 0) {
-			// if empty (e.g. no token)
-			$response = explode(' ', file_get_contents("http://umbrella.jlparry.com/work/registerme/apple/22156b"));
-			$token = $response[1];
+		$current_token = $this->properties->head(1);
+		$token = $current_token[0]->token;
 
-			$db_token = $this->properties->create();
-			$db_token->token = $token;
-			$this->properties->add($db_token);
+		$parts_response = file_get_contents("http://umbrella.jlparry.com/work/mybuilds?key=" . $token);
+		$built_parts = json_decode($parts_response, true);
 
-			$properties = $this->properties->tail();
-		}
+		$response = explode(' ', $parts_response);
 
-		// grab api key (token)
-		foreach ($properties as $token_property) {
-			array_push($tokens, $token_property->token);
-		}
-
-		foreach ($tokens as $token) {
-			$parts_response = file_get_contents("http://umbrella.jlparry.com/work/mybuilds?key=" . $token);
-
-			$built_parts = json_decode($parts_response, true);
-
-			if (sizeof($built_parts) == 0) {
-				continue;
-			}
+		if ($response[0] == '[]') {
+			echo "<b>Not enough parts to retrieve</b>";
+		} else {
+			$size = sizeof($built_parts);
 
 			foreach ($built_parts as $part) {
 				$p = $this->parts->create();
@@ -145,25 +130,31 @@ class PartController extends Application
 				$p->piece = $part['piece'];
 				$p->plant = $part['plant'];
 				$p->stamp = $part['stamp'];
-				$p->piece = $part['piece'];
 				// get line
 				if (preg_match("/^[a-lA-L]$/", $part['model'])) {
 					$p->line = "household";
 				} else if (preg_match("/^[m-vM-V]$/", $part['model'])) {
 					$p->line = "butler";
 				} else {
-					$p-> line = "companion";
+					$p->line = "companion";
 				}
 				$p->isAvailable = 1;
 
 				// insert into database
 				$this->parts->add($p);
 			}
+
+			echo 'Built ' . $size . ' parts';
+
+			$record = array('category' => 'Build', 'description' => 'Built ' . $size . ' parts');
+
+			// Update history table
+			$this->history->add($record);
 		}
 
 		// return to original page
 		$referred_from = $this->session->userdata('referred_from');
-		redirect($referred_from, 'refresh');
+		//redirect($referred_from, 'refresh');
 	}
 
 	/**