+++ application/views/History/history_page.php
@@ -1,38 +1,54 @@
+<script>
+    window.onload = function () {
+        {sort_script}
+        {filterModel_script}
+        {filterSeries_script}
+    };
+</script>
+
 <h1>History View</h1>
 
 <div id="body">
    <!--  transactionId | customerName | date | category | price | description | partId -->
-   <table class="table">
+    {pagination}
+
+    <div class="col-xs-12 text-center">
+        <form method="post" action="">
+            <div class="col-sm-6">
+                <select class="form-control" name="order" id="order">
+                    <option value="stamp">Date</option>
+                    <option value="category">Category</option>
+                </select>
+            </div>
+            <div class="col-sm-6">
+                <input class="btn" type="submit" value="Sort/Filter" />
+            </div>
+        </form>
+    </div>
+    
+    <table class="table">
       <thead class="thead-inverse">
          <tr>
-            <th>Who</th>
-            <th>Category</th>
-            <th>Price</th>
             <th>Date</th>
+            <th>Category</th>
             <th>Description</th>
-            <th>Part ID</th>
+            <th>Amount</th>
          </tr>
       </thead>
       <tbody>
          {history}
          <tr>
             <td>
-                {customer}
+                {date}
             </td>
             <td>
                 {category}
             </td>
             <td>
-                {price}
-            </td>
-            <td>
-                {date}
-            </td>
-            <td>
                 {description}
             </td>
             <td>
-                 {partID}
+              {amount}
             </td>
          </tr>
          {/history}